Back/GreenLandBack/requirements.txt
asgiref==3.8.1
charset-normalizer==3.4.2
Django==5.2.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
flatbuffers==25.2.10
h11==0.16.0
jh2==5.0.9
niquests==3.14.1
openmeteo_requests==1.5.0
openmeteo_sdk==1.20.0
pillow==11.2.1
PyJWT==2.9.0
qh3==1.5.1
sqlparse==0.5.3
typing==3.7.4.3
urllib3-future==2.12.922
wassima==1.2.2